Move response utilities into 'util' directory
